huge refactoring update 2
the following have been completely refactored and are now much neater
and more efficient:
-food & drinks
-reagents/reagent containers
-drinking glasses
-materials
-rnd equipment & designs
-stock parts
-cigarettes and lighters
-fancy boxes
in addition, the following have been added
-crusher
-vehicle UI action buttons
-spraycans/graffiti tagging system
-new reagents addiction system (defaults to nothing, symptoms are
created on a per-reagent basis)
-new reagents overdose system (also defaults to nothing, symptoms are
created on a per-reagent basis)
-digital cameras
-adds new cigarette types from /tg/
-adds joints and rolling paper (but not drying yet, hydroponics refactor
not yet begun)
-shot glasses
-chemistry patches
-police uniforms
-tons of new food and drinks have been added from /tg/
-new arcade game, Orion Trail
-new sound effects for vehicles
-colored glass tables
-prep work for the incoming gun overhaul

reagent and food recipes are currently unchanged, with a refactor
pending. cooking is going to get an overhaul so that will not happen
until then.

tons of path changes for food, maps have been updated accordingly.